Italian COVID-19 data

Datasets containing the italian COVID-19 data.

dpc_covid19_ita_andamento_nazionale

dpc_covid19_ita_regioni

dpc_covid19_ita_province

Format

tibbles with one row per day each levels of detail (regional adn provincial) and as many variables as provided by the original data (see the corresponding sections).

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 440 rows and 24 columns.

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 9240 rows and 30 columns.

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 62998 rows and 14 columns.

Source

https://github.com/pcm-dpc/COVID-19

Details

Licenza: [CC-BY-4.0](https://creativecommons.org/licenses/by/4.0/deed.en) - [Visualizza licenza](https://github.com/pcm-dpc/COVID-19/blob/master/LICENSE)

Editore/Autore del dataset: [Dipartimento della Protezione Civile](http://www.protezionecivile.it/)

Dati forniti dal [Ministero della Salute](http://www.salute.gov.it/)

Elaborazione e gestione dati a cura del [Dipartimento della Protezione Civile](http://www.protezionecivile.it/)

Note

Le Province autonome di Trento e Bolzano sono indicate in "denominazione regione" e con il codice 04 del Trentino Alto Adige.

Ogni Regione ha una Provincia denominata "In fase di definizione/aggiornamento" con il codice provincia da 979 a 999, utile ad indicare i dati ancora non assegnati alle Province.

Data at national level

A tibble with one row per day and 15 variables:

data

Date of notification - Italian time zone)

stato

Country of reference - XYZ (ISO 3166-1 alpha-3)

ricoverati_con_sintomi

Hospitalised patients with symptoms

terapia_intensiva

Intensive Care

totale_ospedalizzati

Total hospitalised patients

isolamento_domiciliare

Home confinement

totale_positivi

Total amount of current positive cases (Hospitalised patients + Home confinement)

variazione_totale_positivi

News amount of current positive cases (Hospitalised patients + Home confinement)

nuovi_positivi

today active cases - yesterday active cases

dimessi_guariti

Recovered

deceduti

Death

casi_da_sospetto_diagnostico

Positive cases emerged from clinical activity

casi_da_screening

Positive cases emerging from surveys and tests, planned at national or regional level

totale_casi

Total amount of positive cases

tamponi

Tests performed

note

Notes

Data at regional level

A tibble with one row per day/region and 19 variables:

data

Date of notification - Italian time zone)

stato

Country of reference - XYZ (ISO 3166-1 alpha-3)

codice_regione

Code of the Region (ISTAT 2019)

denominazione_regione

Name of the Region

lat

Latitude - WGS84

long

Longitude - WGS84

ricoverati_con_sintomi

Hospitalised patients with symptoms

terapia_intensiva

Intensive Care

totale_ospedalizzati

Total hospitalised patients

isolamento_domiciliare

Home confinement

totale_positivi

Total amount of current positive cases (Hospitalised patients + Home confinement)

variazione_totale_positivi

News amount of current positive cases (Hospitalised patients + Home confinement)

nuovi_positivi

Today active cases - yesterday active cases

dimessi_guariti

Recovered

deceduti

Death

casi_da_sospetto_diagnostico

Positive cases emerged from clinical activity

casi_da_screening

Positive cases emerging from surveys and tests, planned at national or regional level

totale_casi

Total amount of positive cases

tamponi

Tests performed

note

Notes

Data at provincial level

A tibble with one row per day/province and 12 variables:

data

Date of notification - Italian time zone)

stato

Country of reference - XYZ (ISO 3166-1 alpha-3)

codice_regione

Code of the Region (ISTAT 2019)

denominazione_regione

Name of the Region

codice_provincia

Code of the Province

denominazione_provincia

Name of the Province

sigla_provincia

Province abbreviation

lat

Latitude - WGS84

long

Longitude - WGS84

totale_casi

Total amount of positive cases

note

Notes