Write RDS for data-raw

Given the names of some data, it save the corresponding objects in the `here::here("data-raw/")` folder. Each object will be saved like an independend `.rds` file with its own name as its filename.

write_raw_rds(data_name)

read_data_raw(data_name)

Arguments

data_name
(chr) names of the pobject

Value

named logical vector (invisibly) with `TRUE`s for all the objects that are correctly saved, `FALSE`s for the other ones.

Functions

read_data_raw: read data-raw